python file comparing works all folders
