+++ results/MoAEpilot_gui_comparisons.txt
@@ -1,3 +1,6 @@
+The Mean Square Error (MSE) calculation of the /home/matay/spmbasics/data/MoAEpilot/sub-01/anat/sub-01_T1w.nii and /home/matay/spmbasics/data/output/MoAEpilot_gui/sub-01/anat/sub-01_T1w.nii, is = 0.0
+The correlation coefficient of the /home/matay/spmbasics/data/MoAEpilot/sub-01/anat/sub-01_T1w.nii and /home/matay/spmbasics/data/output/MoAEpilot_gui/sub-01/anat/sub-01_T1w.nii, is =1.0
+SHA256sums are identical for /home/matay/spmbasics/data/MoAEpilot/sub-01/anat/sub-01_T1w.nii and /home/matay/spmbasics/data/output/MoAEpilot_gui/sub-01/anat/sub-01_T1w.nii
 The Mean Square Error (MSE) calculation of the /home/matay/spmbasics/data/MoAEpilot/sub-01/func/sub-01_task-auditory_bold.nii and /home/matay/spmbasics/data/output/MoAEpilot_gui/sub-01/func/sub-01_task-auditory_bold.nii, is = 0.0
 The correlation coefficient of the /home/matay/spmbasics/data/MoAEpilot/sub-01/func/sub-01_task-auditory_bold.nii and /home/matay/spmbasics/data/output/MoAEpilot_gui/sub-01/func/sub-01_task-auditory_bold.nii, is =1.0
 SHA256sums are identical for /home/matay/spmbasics/data/MoAEpilot/sub-01/func/sub-01_task-auditory_bold.nii and /home/matay/spmbasics/data/output/MoAEpilot_gui/sub-01/func/sub-01_task-auditory_bold.nii
